Bump ORCA version to 3.116 (#11759)
Corresponding ORCA commit: Disallow subquery in predicate of implementation xforms (greenplum-db/gporca#617)

This generates an invalid plan for most cases. While we had this logic
in some xforms, we missed a few cases.

This manifested as a crash in some cases as we assumed subqueries had
been handled by time we get to optimization stage.

Additionally, add error handling to PexprScalarExactChild()

This function is only called in places where the exact child must exist,
and returns null otherwise.

However, there have been some cases where we assume the child must
exist, but due to another bug the child does not and we don't handle the
null. This is a defensive change; if we get into this situation we'll
throw an exception and fall back to planner instead of crashing.