enhance OLTP performance by skipping prepare phase for read only segments
Previously, if a distributed transaction has 1 segment wrote WAL and 2
segments just executed read only queries, 2PC will be used for this
distributed transaction. But actually, 1PC is sufficient. For example,
assuming there is a table test1(c1 int, c2 int) distributed by c1, and
it has three tuples (1,2),(2,3),(3,4). If we have a dtx like this:
    BEGIN;
    update test1 set c2 = 1 where c1 = 1;
    select * from test1;
    COMMIT;
The dtx will write xlog in a segment, while other segments will just
read the table data, so 1PC should be used, but without this commit
2PC will be used.

So this commit records segments that has written WAL and segments that
are read only during the DTX into two separate lists. For those read
only segments we skip prepare phase, and for the others 1PC or
2PC is used based on its number. Besides, for replicated table, we assign
it to the segment that is already in the above two lists, to reduce the potential
overhead of creating gang and commit message. With the optimizations
mentioned above, we can reduce the dtx commit message overhead, and OLTP
performance can be improved as a result.

A typical case for this optimization is TPC-C benchmark workload.
In standard TPC-C benchmark, 45% of its transaction is newOrder, and
in the newOrder transaction, item table (only used by newOrder) won't
be modified, moreover, 90% transactions of newOrder will only access data of
a given warehouse id, so we can create item table as a replicated table,
and with this commit, 90% newOrder transactions will use 1PC instead of 2PC,
which can help improve the throughput of TPC-C workload.

Another benefit is that it can avoid to assign local transaction id in read only segments,
which can help decrease the needs of doing vacuum freeze in segments.