Drop PostmasterContext after InitPostgres()
PostmasterContext is not necessary for a normal postgres backend
process, drop it to save memory.

In a corner case, the flat files it loaded into memory, like hba, ident
and role files, could be big, which would be a serious issue if the
postmaster forks them with thousands of copies without vmem tracking.

FYI, CoW memory is counted by system OOM killer if the Linux
configuration vm.overcommit_memory is 2.