Coerce unknown-type literals to type cstring, but hide it
Commit 0cbadfa0242 "Coerce unknown-type literals to type text instead of
cstring" has an issue that it processed outfunc and infunc directly,
which made Greenplum have no chance to do an unknown-to-text-to-cstring
cast. It errors out, even panics, in some cases since the data type
don't match.

The better solution is still coercing unknown-type literals to type
cstring, but hiding it. Check out the test case to see what it fixes.

Reviewed-by: Jesse Zhang <sbjesse@gmail.com>
Reviewed-by: Ashwin Agrawal <aashwin@vmware.com>
Reviewed-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>