Use the correct flow in ParallelizeSubplan for subplan (#12457)
When creating a plan with subplan in ParallelizeSubplan,it use
the wrong flow to decide motion type for subplan. And the wrong
flow will cause subplan set wrong motionType and cause assertion
failure Assert(recvSlice->gangSize == 1) in nodeMotion.c.

When ParallelizeSubplan(), if the subplan is uncorrelated, multi-row
subquery,then it either focuses or broadcasts the subplan based on the
flow which describe the containing plan node's slice execution position.
Actually the flow should be the top-level flow for the corresponding slice
instead of the containing plan node's flow. To select the correct flow during
the plan tree iteration, we only set currentPlanFlow when jump into a new slice.

This fixed issue: #12371