FileRepResync_UpdateEntry:Use updated entry to log
We were using the passed in entry in the log statement. Use the updated
entry instead, as the passed in entry might not have all of its fields
populated - as is the case for the callsite in
FileRepPrimary_ResyncBufferPoolIncrementalWrite().

Also, in case the lookup returned NULL, we used to log as if nothing
happened. So, we add an appropriate WARNING here.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>