Impove the performance for segment list updating.
Before querying the bitmap in addToGxactDtxSegments() and
addToGxactReadOnlySegments(), we should not switch the memory context to
TopTransactionContext, which would reduce the overhead caused by memory context
switching.

In currentDtxDispatchProtocolCommand(), the complexity of
list_concat_unique_int() is O(n^2), we use foreach and lappend_int() to
impove the complexity to O(n).

Co-authored-by: Xing Guo <higuoxing@gmail.com>
Co-authored-by: jingwen-yang-yjw <jingweny1@vmware.com>
Co-authored-by: Yongtao Huang <yongtaoh@vmware.com>